Remove tabs and fix spacing: ipv6 scripts
The network-scripts are a mix of tabs and spaces (!!) as well as
completely inconsistent indentation (some 2, some 4).

This is a series of 3 patches to move the entire of network-scripts to
2-space indentation for consistency so I can develop on the codebase
without my brain bleeding. I don't feel strongly about 2 vs 4... 2 was
just most common in the first file I picked, so I used that.

This set is in 3 sets for easier reviewing and easier reverting if
necessary:

* 1/3 - ipv6 files
* 2/3 - network-functions - it needed it's own
* 3/3 - ifup/ifdown files